address comments && cleanup
