+++ backend-server/src/api/controllers/interviewerController.ts
@@ -13,6 +13,7 @@ export async function addInterviewer(
     email,
     name,
   };
+
   await dataAccess.setInterviewer(interviewer);
 }
