+++ backend-server/src/api/routes/availability.ts
@@ -1,32 +1,33 @@
 import express from "express";
 import {
   addAvailability,
-  getAvailability,
   AddAvailabilityBody,
   GetAvailabilityParams,
-  getAllAvailabilities,
+  getInterviewerAvailabilities,
   ReplaceAvailabilitiesBody,
   getAllCalendarAvailabilities,
   replaceAllAvailabilities,
   GetMergedRoutesParams,
 } from "../controllers/availabilityController";
-import { findOverlapping } from "../controllers/merge";
+import { findOverlapping } from "../controllers/mergeController";
 import { Availability } from "../data/models";
 
 export const availabilityRouter = express.Router();
 
 availabilityRouter.post("/", async (req, res) => {
+  const body: AddAvailabilityBody = {
+    organization: req.body.organization,
+    interviewerUID: req.body.interviewerUID,
+    startTime: req.body.startTime,
+    isBooked: req.body.isBooked,
+    bookedByEmail: req.body.bookedByEmail,
+    durationMins: req.body.durationMins,
+  };
+
   try {
-    const body: AddAvailabilityBody = {
-      organization: req.body.organization,
-      interviewerUID: req.body.interviewerUID,
-      startTime: req.body.startTime,
-      isBooked: req.body.isBooked,
-      bookedByEmail: req.body.bookedByEmail,
-      durationMins: req.body.durationMins,
-    };
     if (!Object.values(body).every((field) => field != null))
       throw new Error(`Incomplete Request Body: ${JSON.stringify(body)}`);
+
     await addAvailability(body);
     res.send(
       `A timeslot at ${body.startTime} has been added to interviewer ${body.interviewerUID}'s availability`
@@ -37,14 +38,16 @@ availabilityRouter.post("/", async (req, res) => {
 });
 
 availabilityRouter.put("/", async (req, res) => {
+  const body: ReplaceAvailabilitiesBody = {
+    eventsAPI: req.body.eventsAPI,
+    interviewerUID: req.body.interviewerUID,
+    organization: req.body.organization,
+  };
+
   try {
-    const body: ReplaceAvailabilitiesBody = {
-      eventsAPI: req.body.eventsAPI,
-      interviewerUID: req.body.interviewerUID,
-      organization: req.body.organization,
-    };
     if (!Object.values(body).every((field) => field != null))
       throw new Error(`Incomplete Request Body: ${JSON.stringify(body)}`);
+
     const createdAvailabilities = await replaceAllAvailabilities(body);
     res.send(
       `${createdAvailabilities.length} timeslots have been added to ${body.interviewerUID}'s availabilities`
@@ -55,30 +58,15 @@ availabilityRouter.put("/", async (req, res) => {
 });
 
 availabilityRouter.get("/", async (req, res) => {
-  try {
-    const body: GetAvailabilityParams = {
-      organization: req.query.organization as string,
-      interviewerUID: req.query.interviewerUID as string,
-      startTime: req.query.startTime as string,
-    };
-    if (!Object.values(body).every((field) => field != null))
-      throw new Error(`Incomplete Request Body: ${JSON.stringify(body)}`);
-    const availabilityData = await getAvailability(
-      body.organization,
-      body.interviewerUID,
-      body.startTime
-    );
-    res.send(availabilityData);
-  } catch (err) {
-    res.status(500).send(`error processing request: ${err}`);
-  }
-});
-
-availabilityRouter.get("/", async (req, res) => {
   const organization = req.query.organization as string;
   const interviewerUID = req.query.interviewerUID as string;
+  const body: GetAvailabilityParams = { organization, interviewerUID };
+
+  if (!Object.values(body).every((field) => field != null))
+    throw new Error(`Incomplete Request Body: ${JSON.stringify(body)}`);
+
   try {
-    const availabilityData = await getAllAvailabilities(
+    const availabilityData = await getInterviewerAvailabilities(
       organization,
       interviewerUID
     );
@@ -91,6 +79,7 @@ availabilityRouter.get("/", async (req, res) => {
 availabilityRouter.get("/calendarAvailabilities", async (req, res) => {
   const organization = req.query.organization as string;
   const interviewerUID = req.query.interviewerUID as string;
+
   try {
     const calendarAvailabilitiesData = await getAllCalendarAvailabilities(
       organization,
@@ -108,22 +97,22 @@ availabilityRouter.get("/mergedTimes", async (req, res) => {
     interviewerUID1: req.query.interviewerUID1 as string,
     interviewerUID2: req.query.interviewerUID2 as string,
   };
+
   if (!Object.values(body).every((field) => field != null))
     throw new Error(`Incomplete Request Body: ${JSON.stringify(body)}`);
 
   try {
-    const allAvailabilities1 = (await getAllAvailabilities(
+    const allAvailabilities1 = (await getInterviewerAvailabilities(
       body.organization,
       body.interviewerUID1
     )) as Availability[];
 
-    const allAvailabilities2 = (await getAllAvailabilities(
+    const allAvailabilities2 = (await getInterviewerAvailabilities(
       body.organization,
       body.interviewerUID2
     )) as Availability[];
 
     const merged = findOverlapping(allAvailabilities1, allAvailabilities2);
-
     res.send(merged);
   } catch (err) {
     res.send(`error processing request: ${err}`);
