+++ backend-server/src/api/routes/interviewer.ts
@@ -14,6 +14,7 @@ interviewerRouter.post("/", async (req, res) => {
     email,
     name,
   }: Interviewer = req.body;
+
   try {
     await addInterviewer(organization, userUID, email, name);
     res.send(`${userUID} has been added to ${organization}`);
@@ -25,6 +26,7 @@ interviewerRouter.post("/", async (req, res) => {
 interviewerRouter.get("/", async (req, res) => {
   const organization = req.query.organization as string;
   const userUID = req.query.userUID as string;
+
   try {
     const interviewerData = await getInterviewer(organization, userUID);
     res.send(interviewerData);
