+++ backend-server/src/api/versions/v1.ts
@@ -8,4 +8,4 @@ export const v1Router = express.Router();
 
 v1Router.use("/interviewers", interviewerRouter);
 v1Router.use("/availabilities", availabilityRouter);
-v1Router.use("/event", eventRouter);
+v1Router.use("/events", eventRouter);
